The content of PushButton's and ToggleButton's now shifts down and to the right 1px when click, indicating that the button has been pressed.  Without this style change, it isn't obvious that the button's state is changing.
Menu menu selection in the Showcase has been improved so that the examples display synchronously when a menu item is clicked.  Previously, we used a history token to trigger a new example, but this could introduce a delay up to 250ms in some browsers.

Finally, the a new StyleSheetLoader class has been introduced that reliably detects when a style sheet has loaded.  We use this when switching between themes so we can handle various browser-specific rendering issues after the new theme has loaded.  To support this, each GWT theme style sheet now includes a unique style definition that can be used to detect when the style sheet has loaded (see the comments in the .css files for more info). 

Patch by: jlabanca
Review by: ecc (desk review)



git-svn-id: http://google-web-toolkit.googlecode.com/svn/releases/1.5@2997 8db76d5a-ed1c-0410-87a9-c151d255dfc7