Addresses some of the remaining issues from
http://groups.google.com/group/Google-Web-Toolkit-Contributors/browse_thread/thread/486babc9bb0863a9

In particular, nukes the unused Iterator class from ElementMapper,
fixes a bug in DatePicker#removeStyleFromVisibleDates, touches up some
javadoc, removes unneeded redeclaration of refresh from CalendarView.

We still lack unit tests on CalendarView and ElementMapper, and time
may not permit them. There is also a discussion in progress on
removeStyleFromVisibleDates.  When that is done, any remaining methods
with Iterator in their signature will need to re-worked to use
Collection instead.



git-svn-id: http://google-web-toolkit.googlecode.com/svn/branches/1_6_datepicker@4239 8db76d5a-ed1c-0410-87a9-c151d255dfc7