Renames some methods to reduce the confusion between global and not global styles.  Also fixes various varargs methods, which weren't really working.
We now have styles (formerly global styles) and transient styles, the
latter of which can only be applied to visible dates. The transient word
appears in the former setEnabled method as well, for consistency.

Removes separate methods for removing global and transient styles, as
the former method would always remove the transient styles anyway, and
it seems very unlikely that a client will use the same style name for
both a transient and permanent date style.

Adds unit tests for styles and enabling.

TBR: ecc
submitter: rjrjr



git-svn-id: http://google-web-toolkit.googlecode.com/svn/branches/1_6_datepicker@4245 8db76d5a-ed1c-0410-87a9-c151d255dfc7