This patch adds a modular testing framework for adding ApiChecker tests. In addition, it fixes the following 3 cases where the ApiChecker was wrong and adds tests for each case.
(i) A api class is made FINAL and simultaneously the FINAL keyword is
removed from its methods.

(ii) An api class's methods are overloaded and the overloading is such that the
methods are "compatible."

(iii) A api class's fields are moved to its superclass. 


Patch by: amitmanjhi
Review by: ecc (desk review)



git-svn-id: http://google-web-toolkit.googlecode.com/svn/releases/1.6@4514 8db76d5a-ed1c-0410-87a9-c151d255dfc7