This change restores TextBoxBase's implementation of HasChangeEventHandlers
We recently removed TextBox's implementation of HasChangeEventHandlers,
reasoning that it was redundant with its implementation of
HasValueChange<String>. 

As I struggle to integrate this change with a ton of existing code that has
already sprung up relying on TextBox#addChangeHandler, in many cases composites
rethrowing the ChangeEvent, it's becoming clear that this was a mistake. 

As it stands now, TextBoxBase hides the dom event that initiated a change. This
is inconsistent with how ButtonBase propagates dom click events, and is a loss
of functionality from the old event system. Just because our widgets "adds
value" with the logical ValueChangeEvents doesn't mean we should be trying to
hide the dom events behind them.

reviewer: jlabanca
testing: ant test on linux



git-svn-id: http://google-web-toolkit.googlecode.com/svn/releases/1.6@4583 8db76d5a-ed1c-0410-87a9-c151d255dfc7